app/requirements.txt
Flask
mysql-connector
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy